Change Response to be Fully JSON - New Branch Due to Test Issues
While testing, I realized that there was no reason to have multiple instances of json.Marshal
throughout my program. Not only that, but having these extras was causing
objects to be double-stringified, making the results a lot harder
to use for anyone writing a program that interfaces with the API'

Starting this change as a new pull request since I have a lot to figure out with the testing now.