Clear up Redis logging issues, add returns for errors in handler functions
Formatted Redis logging to be a more usable JSON string, as well as include
a timestamp so that each log is unique even if it's the same request.

Also added returns in the handlers so that when an error occurs
it does not keep executing the rest of the code.